fix(api): dist import error
