+++ packages/api/src/trpc.ts
@@ -1,10 +1,2 @@
-import { initTRPC } from '@trpc/server'
-import type { Context } from '.'
-import { authedMiddleware } from './middleware/auth'
-
-export const t = initTRPC.context<Context>().create()
-
-export const router = t.router
-export const publicProcedure = t.procedure
-export const authedProcedure = t.procedure.use(authedMiddleware)
-export const createCallerFactory = t.createCallerFactory
+export { t, router, publicProcedure, createCallerFactory } from './trpc/core'
+export { authedMiddleware, authedProcedure } from './trpc/middlewares/authed'
