+++ packages/api/src/trpc/context.ts
@@ -1,0 +1,10 @@
+import { getAuth } from '@foodwise/auth/server'
+import { db } from '@foodwise/db/client'
+
+export const createTRPCContext = async ({ headers }: { headers: Headers }) => ({
+  headers,
+  session: await getAuth().api.getSession({ headers }),
+  db,
+})
+
+export type Context = Awaited<ReturnType<typeof createTRPCContext>>
