+++ packages/api/src/trpc/core.ts
@@ -1,0 +1,9 @@
+import { initTRPC } from '@trpc/server'
+import type { Context } from './context'
+
+
+export const t = initTRPC.context<Context>().create()
+
+export const router = t.router
+export const publicProcedure = t.procedure
+export const createCallerFactory = t.createCallerFactory
